Put modulus operator in shift.
The issue is that a particle could, in principle, jump across the entire 
domain multiple times in one timestep, if it were a fast-moving particle 
and it suffered no collisions.

Therefore, when particles are shifted to be back in the main domain 
(-L/2,L/2) say, there really should be a modulus operation performed on 
the sheared velocity to make sure it reflects the number of times the 
particle potentially crossed through the reflected domain.